+++ src/integration-test/java/uk/gov/ida/integrationtest/helpers/AssertionHelper.java
@@ -1,5 +1,7 @@
 package uk.gov.ida.integrationtest.helpers;
 
+import com.google.common.collect.ImmutableList;
+import org.apache.commons.codec.binary.Base64;
 import org.joda.time.DateTime;
 import org.opensaml.saml.saml2.core.Assertion;
 import org.opensaml.saml.saml2.core.Attribute;
@@ -14,29 +16,35 @@ import org.opensaml.saml.saml2.core.SubjectConfirmationData;
 import org.opensaml.saml.saml2.core.impl.AudienceBuilder;
 import org.opensaml.saml.saml2.core.impl.AudienceRestrictionBuilder;
 import org.opensaml.saml.saml2.core.impl.ConditionsBuilder;
+import org.opensaml.saml.saml2.encryption.Decrypter;
 import org.opensaml.xmlsec.signature.Signature;
+import uk.gov.ida.common.shared.security.PrivateKeyFactory;
+import uk.gov.ida.common.shared.security.PublicKeyFactory;
+import uk.gov.ida.common.shared.security.X509CertificateFactory;
 import uk.gov.ida.saml.core.extensions.IdaAuthnContext;
 import uk.gov.ida.saml.core.test.TestCredentialFactory;
+import uk.gov.ida.saml.core.test.builders.AddressAttributeBuilder_1_1;
+import uk.gov.ida.saml.core.test.builders.AddressAttributeValueBuilder_1_1;
 import uk.gov.ida.saml.core.test.builders.AttributeQueryBuilder;
 import uk.gov.ida.saml.core.test.builders.AuthnStatementBuilder;
+import uk.gov.ida.saml.core.test.builders.DateAttributeBuilder_1_1;
+import uk.gov.ida.saml.core.test.builders.GenderAttributeBuilder_1_1;
 import uk.gov.ida.saml.core.test.builders.SubjectConfirmationBuilder;
 import uk.gov.ida.saml.core.test.builders.SubjectConfirmationDataBuilder;
+import uk.gov.ida.saml.security.AssertionDecrypter;
+import uk.gov.ida.saml.security.DecrypterFactory;
+import uk.gov.ida.saml.security.IdaKeyStore;
+import uk.gov.ida.saml.security.IdaKeyStoreCredentialRetriever;
+import uk.gov.ida.saml.security.validators.encryptedelementtype.EncryptionAlgorithmValidator;
 
+import java.security.KeyPair;
+import java.security.PrivateKey;
+import java.security.PublicKey;
 import java.util.List;
 
 import static java.util.Arrays.asList;
 import static java.util.Collections.singletonList;
-import static uk.gov.ida.matchingserviceadapter.services.AttributeQueryServiceTest.anEidasSignature;
-import static uk.gov.ida.saml.core.test.TestCertificateStrings.HUB_TEST_PRIVATE_SIGNING_KEY;
-import static uk.gov.ida.saml.core.test.TestCertificateStrings.HUB_TEST_PUBLIC_SIGNING_CERT;
-import static uk.gov.ida.saml.core.test.TestCertificateStrings.STUB_COUNTRY_PUBLIC_PRIMARY_CERT;
-import static uk.gov.ida.saml.core.test.TestCertificateStrings.STUB_COUNTRY_PUBLIC_PRIMARY_PRIVATE_KEY;
-import static uk.gov.ida.saml.core.test.TestCertificateStrings.STUB_IDP_PUBLIC_PRIMARY_CERT;
-import static uk.gov.ida.saml.core.test.TestCertificateStrings.STUB_IDP_PUBLIC_PRIMARY_PRIVATE_KEY;
-import static uk.gov.ida.saml.core.test.TestCertificateStrings.TEST_RP_MS_PRIVATE_ENCRYPTION_KEY;
-import static uk.gov.ida.saml.core.test.TestCertificateStrings.TEST_RP_MS_PUBLIC_ENCRYPTION_CERT;
-import static uk.gov.ida.saml.core.test.TestCertificateStrings.TEST_RP_PRIVATE_SIGNING_KEY;
-import static uk.gov.ida.saml.core.test.TestCertificateStrings.TEST_RP_PUBLIC_SIGNING_CERT;
+import static uk.gov.ida.saml.core.test.TestCertificateStrings.*;
 import static uk.gov.ida.saml.core.test.TestEntityIds.HUB_ENTITY_ID;
 import static uk.gov.ida.saml.core.test.TestEntityIds.HUB_SECONDARY_ENTITY_ID;
 import static uk.gov.ida.saml.core.test.TestEntityIds.STUB_IDP_ONE;
@@ -50,6 +58,8 @@ import static uk.gov.ida.saml.core.test.builders.AuthnStatementBuilder.anEidasAu
 import static uk.gov.ida.saml.core.test.builders.IPAddressAttributeBuilder.anIPAddress;
 import static uk.gov.ida.saml.core.test.builders.IssuerBuilder.anIssuer;
 import static uk.gov.ida.saml.core.test.builders.NameIdBuilder.aNameId;
+import static uk.gov.ida.saml.core.test.builders.PersonNameAttributeBuilder_1_1.aPersonName_1_1;
+import static uk.gov.ida.saml.core.test.builders.PersonNameAttributeValueBuilder.aPersonNameValue;
 import static uk.gov.ida.saml.core.test.builders.SignatureBuilder.aSignature;
 import static uk.gov.ida.saml.core.test.builders.SimpleStringAttributeBuilder.aSimpleStringAttribute;
 import static uk.gov.ida.saml.core.test.builders.SubjectBuilder.aSubject;
@@ -70,6 +80,46 @@ public class AssertionHelper {
 
     private static Signature aValidSignature() {
         return aValidSignature(STUB_IDP_PUBLIC_PRIMARY_CERT, STUB_IDP_PUBLIC_PRIMARY_PRIVATE_KEY);
+    }
+
+    public static AssertionDecrypter anAssertionDecrypter() {
+        PublicKeyFactory publicKeyFactory = new PublicKeyFactory(new X509CertificateFactory());
+        PrivateKey privateKey = new PrivateKeyFactory().createPrivateKey(Base64.decodeBase64(PRIVATE_SIGNING_KEYS.get(HUB_ENTITY_ID)));
+        PublicKey publicKey = publicKeyFactory.createPublicKey(getPrimaryPublicEncryptionCert(HUB_ENTITY_ID));
+
+        PrivateKey privateEncryptionKey = new PrivateKeyFactory().createPrivateKey(Base64.decodeBase64(HUB_TEST_PRIVATE_ENCRYPTION_KEY));
+        PublicKey publicEncryptionKey = publicKeyFactory.createPublicKey(HUB_TEST_PUBLIC_ENCRYPTION_CERT);
+
+        KeyPair encryptionKeyPair = new KeyPair(publicEncryptionKey, privateEncryptionKey);
+
+        KeyPair signingKeyPair = new KeyPair(publicKey, privateKey);
+        IdaKeyStore keyStore = new IdaKeyStore(signingKeyPair, singletonList(encryptionKeyPair));
+
+        IdaKeyStoreCredentialRetriever idaKeyStoreCredentialRetriever = new IdaKeyStoreCredentialRetriever(keyStore);
+        Decrypter decrypter = new DecrypterFactory().createDecrypter(idaKeyStoreCredentialRetriever.getDecryptingCredentials());
+        return new AssertionDecrypter(new EncryptionAlgorithmValidator(), decrypter);
+    }
+
+
+    public static Assertion aCompleteMatchingDatasetAssertion(String requestId) {
+        List<Attribute> attributes = asList(
+            aPersonName_1_1().addValue(aPersonNameValue().withValue("OldSurname").withFrom(new DateTime(1990, 1, 30, 0, 0)).withTo(new DateTime(2000, 1, 29, 0, 0)).withVerified(true).build()).buildAsSurname(),
+            aPersonName_1_1().addValue(aPersonNameValue().withValue("CurrentSurname").withVerified(true).build()).buildAsSurname(),
+            aPersonName_1_1().addValue(aPersonNameValue().withValue("FirstName").withVerified(false).build()).buildAsFirstname(),
+            AddressAttributeBuilder_1_1.anAddressAttribute().addAddress(new AddressAttributeValueBuilder_1_1().addLines(ImmutableList.of("address line 1")).withVerified(false).build()).buildCurrentAddress(),
+            AddressAttributeBuilder_1_1.anAddressAttribute().addAddress(new AddressAttributeValueBuilder_1_1().addLines(ImmutableList.of("address line 2")).withVerified(true).build()).buildPreviousAddress(),
+            GenderAttributeBuilder_1_1.aGender_1_1().build(),
+            DateAttributeBuilder_1_1.aDate_1_1().buildAsDateOfBirth());
+
+        return aMatchingDatasetAssertion(attributes, false, requestId);
+    }
+
+    public static Assertion assertionWithOnlyFirstName(String requestId) {
+        List<Attribute> attributes = asList(
+            aPersonName_1_1().addValue(aPersonNameValue().withValue("SteveFirstName").withFrom(new DateTime(2000, 1, 30, 0, 0)).withTo(new DateTime(2010, 12, 30, 0, 0)).build()).buildAsFirstname(),
+            aPersonName_1_1().addValue(aPersonNameValue().withValue("Surname").build()).buildAsSurname());
+
+        return aMatchingDatasetAssertion(attributes, false, requestId);
     }
 
     public static Assertion anAuthnStatementAssertion(String inResponseTo) {
